+++ Modules/Utils/ssmAlignmentOptions.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "ssmBaseOptions.h"
+
+namespace ssm
+{
+
+}
